+++ packages/core/package.json
@@ -27,6 +27,7 @@
     "@babel/plugin-proposal-class-properties": "^7.1.0",
     "@babel/preset-env": "^7.1.0",
     "@babel/preset-react": "^7.0.0",
+    "enzyme": "^3.7.0",
     "react": "^16.5.2"
   },
   "babel": {
